Initial Commit
Included all files for final prototype version of project (excluding video demonstration and FAA data excel file due to size issue). Designed to run in PyCharm and on Windows 10 PowerShell using the following libraries for the user interface: webbrowser, ctypes, html, time, os and for the heatmap: pandas, folium, plotly_express, and openpyxl. 

All text documents can be considered to be previous versions of the prototype, as they contain old code that was documented at various times during the three weeks of the prototype making process. All sources are documented on the interface or otherwise in code comments. 

Special thanks to Professor Hannah Vander Zanden, the 2024 class of 'Is Animal Migration Disappearing?' Quest course, Adam Gulde of Open Source Club, Teacher Assistant Lauren Escarcha of Data Structures and Algorithms, Shan Sundal, and to family and friends that made this project possible.